pyproject.toml
[tool.poetry]
name = "gamerzrpg"
version = "0.1.0"
description = ""
authors = ["Navi"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiosqlite = "^0.17.0"
"discord.py" = "^2.0.1"
python-dotenv = "^0.21.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
